Rename visit to visitor
