+++ frontend/src/metabase-lib/v1/expressions/diagnostics/check-arg-validators.ts
@@ -3,7 +3,7 @@ import * as Lib from "metabase-lib";
 import { getClauseDefinition } from "../config";
 import { DiagnosticError } from "../errors";
 import { getToken } from "../utils";
-import { visit } from "../visit";
+import { visit } from "../visitor";
 
 export function checkArgValidators({
   expressionParts,
